Rename runtime/Mem.h to Allocators.h
In advance of flattening the namespace galois::runtime into galois,
rename galois/runtime/Mem.h to avoid clobbering galois/Mem.h

Allocators themselves are more an implementation detail of Mem.h so
prefer users to include Mem.h over Allocators.h.
